Addressing PR feedback
